Restructured for distribution
