+++ cli.py
@@ -1,0 +1,6 @@
+from fotocopy.__main__ import main
+
+if __name__ == "__main__":
+	main()
+	
+
