+++ fotocopy/fotoexif.py
