+++ fotocopy/fotoinfo.py
